+++ changelog.md
@@ -1,8 +1,12 @@
+### 2.2.10.0
+- Improved start-up service detection speed
+- Adding the auto update Palace.AutoUpdate plugin to nuget
+
 ### 2.0.8.0
-Split Autoupdate to Palace.AutoUpdate assembly
+- Split Autoupdate to Palace.AutoUpdate assembly
 
 ### 1.3.6.0
-Add mandatory Initialize method for poco services
+- Add mandatory Initialize method for poco services
 
 ### 1.0.0.0
-Start
+- Start